Modifications to when to serialize null values
In order to conform with the API documentation, I've
made some changes to the default behaviour when sending
requests using SignedAPI.java.

This changes the default behaviour of always serializing
nulls to only doing so in the case of sending a request
with the VIEW_AUTOMATIC_SETTLEMENT_DETAILS_CSV method.
For this method, a null value is required by the API
for the attributes parameter.

For all other requests, null values can be omitted and
this commit makes it so.