RecipientInformation: Add toString method
When creating the signature from a Deposit object
containing a RecipientInformation object, the plain
string of the signature contained not a serialized
representation of the RI object, but instead its
object address given by Object.toString.

This commit overrides the default toString in
order to make a serializable representation of
the RecipientInformation object.