RecipientInformation: Removed toString, now extends AttributeData class
Instead of manually serializing the object using a toString method,
I want to rely on our already existing serializeData method in the
SignatureHandler class.

In order to prepare for this fix, I've made a new class called
AttributeData, which is to be used for signaling to the
SignatureHandler class that this object needs to be serialized
and not use toString.

I removed the tests for toString since they no longer have
any purpose. I have however left the Junit 4 dependency in the
POM for future tests.