Disregard null values for AttributeData
When generating the request signature, we do not
wish to include optional values that are set to
null as they can be omitted.

An oversight led to us still including fields with
null values for the RecipientInformation class.

With this commit, we now ignore fields with null
values for instances of Attribute data as well.